Move asset discovery to kernel filesystem
